requirements.txt
Package                       Version
----------------------------- ---------
anyio                         3.3.0
argon2-cffi                   20.1.0
async-generator               1.10
attrs                         21.2.0
Babel                         2.9.1
backcall                      0.2.0
bleach                        4.0.0
cached-property               1.5.2
cachetools                    4.2.2
certifi                       2021.5.30
cffi                          1.14.6
charset-normalizer            2.0.4
colorama                      0.4.4
cycler                        0.10.0
debugpy                       1.4.1
decorator                     5.0.9
defusedxml                    0.7.1
entrypoints                   0.3
et-xmlfile                    1.1.0
google-api-core               1.31.1
google-auth                   1.34.0
google-cloud-bigquery         2.23.2
google-cloud-bigquery-storage 2.6.2
google-cloud-core             1.7.2
google-cloud-datastore        2.1.6
google-cloud-storage          1.41.1
google-crc32c                 1.1.2
google-resumable-media        1.3.3
googleapis-common-protos      1.53.0
grpcio                        1.39.0
h5py                          3.3.0
idna                          3.2
importlib-metadata            3.10.1
ipykernel                     6.0.3
ipython                       7.26.0
ipython-genutils              0.2.0
ipywidgets                    7.6.3
jedi                          0.18.0
Jinja2                        3.0.1
joblib                        1.0.1
json5                         0.9.6
jsonschema                    3.2.0
jupyter-client                6.1.12
jupyter-core                  4.7.1
jupyter-server                1.10.2
jupyterlab                    3.1.1
jupyterlab-pygments           0.1.2
jupyterlab-server             2.6.2
jupyterlab-widgets            1.0.0
Keras                         2.4.3
kiwisolver                    1.3.1
libcst                        0.3.19
MarkupSafe                    2.0.1
matplotlib                    3.4.2
matplotlib-inline             0.1.2
mistune                       0.8.4
mypy-extensions               0.4.3
nbclassic                     0.3.1
nbclient                      0.5.3
nbconvert                     6.1.0
nbformat                      5.1.3
nest-asyncio                  1.5.1
notebook                      6.4.0
numpy                         1.21.1
openpyxl                      3.0.7
packaging                     21.0
pandas                        1.3.1
pandocfilters                 1.4.3
parso                         0.8.2
pickleshare                   0.7.5
Pillow                        8.3.1
pip                           21.2.2
prometheus-client             0.11.0
prompt-toolkit                3.0.19
proto-plus                    1.19.0
protobuf                      3.17.3
pyarrow                       5.0.0
pyasn1                        0.4.8
pyasn1-modules                0.2.8
pycparser                     2.20
Pygments                      2.9.0
pyparsing                     2.4.7
pyrsistent                    0.18.0
python-dateutil               2.8.2
pytz                          2021.1
pywin32                       301
pywinpty                      1.1.3
PyYAML                        5.4.1
pyzmq                         22.1.0
requests                      2.26.0
requests-unixsocket           0.2.0
rsa                           4.7.2
scikit-learn                  0.24.2
scipy                         1.7.1
seaborn                       0.11.1
Send2Trash                    1.7.1
setuptools                    57.4.0
six                           1.16.0
sklearn                       0.0
sniffio                       1.2.0
terminado                     0.10.1
testpath                      0.5.0
threadpoolctl                 2.2.0
tornado                       6.1
traitlets                     5.0.5
typing-extensions             3.10.0.0
typing-inspect                0.7.1
urllib3                       1.26.6
wcwidth                       0.2.5
webencodings                  0.5.1
websocket-client              1.1.0
widgetsnbextension            3.5.1
zipp                          3.5.0
anyio==3.3.0
argon2-cffi==20.1.0
async-generator==1.10
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bleach==4.0.0
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
colorama==0.4.4
cycler==0.10.0
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
entrypoints==0.3
et-xmlfile==1.1.0
google-api-core==1.31.1
google-auth==1.34.0
google-cloud-bigquery==2.23.2
google-cloud-bigquery-storage==2.6.2
google-cloud-core==1.7.2
google-cloud-datastore==2.1.6
google-cloud-storage==1.41.1
google-crc32c==1.1.2
google-resumable-media==1.3.3
googleapis-common-protos==1.53.0
grpcio==1.39.0
h5py==3.3.0
idna==3.2
importlib-metadata==3.10.1
ipykernel==6.0.3
ipython==7.26.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
json5==0.9.6
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-server==1.10.2
jupyterlab==3.1.1
jupyterlab-pygments==0.1.2
jupyterlab-server==2.6.2
jupyterlab-widgets==1.0.0
Keras==2.4.3
kiwisolver==1.3.1
libcst==0.3.19
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
mypy-extensions==0.4.3
nbclassic==0.3.1
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.21.1
openpyxl==3.0.7
packaging==21.0
pandas==1.3.1
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.1
prometheus-client==0.11.0
prompt-toolkit==3.0.19
proto-plus==1.19.0
protobuf==3.17.3
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.9.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
pywin32==301
pywinpty==1.1.3
PyYAML==5.4.1
pyzmq==22.1.0
requests==2.26.0
requests-unixsocket==0.2.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.7.1
seaborn==0.11.1
Send2Trash==1.7.1
six==1.16.0
sklearn==0.0
sniffio==1.2.0
terminado==0.10.1
testpath==0.5.0
threadpoolctl==2.2.0
torch==1.9.0
torchaudio==0.9.0
torchvision==0.10.0
tornado==6.1
traitlets==5.0.5
typing-extensions==3.10.0.0
typing-inspect==0.7.1
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.1.0
widgetsnbextension==3.5.1
zipp==3.5.0

swifter~=1.0.9
tqdm~=4.62.0